requirements.txt
fastapi
uvicorn
python-dotenv
python-multipart
pymupdf
google-generativeai
sentence-transformers
faiss-cpu
openai
pytesseract 
pillow
python-docx
pandas
openpyxl
python-dateutil
odfpy
